Rework common controller
