+++ pkg/utils/common/deployment.go
@@ -1,0 +1,44 @@
+package common
+
+import (
+	"context"
+
+	appsv1 "k8s.io/api/apps/v1"
+	k8sErrors "k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/wait"
+)
+
+// GetAppDeploymentByName returns the deployment for a given component name
+func (h *SuiteController) GetAppDeploymentByName(deploymentName string, deploymentNamespace string) (*appsv1.Deployment, error) {
+	namespacedName := types.NamespacedName{
+		Name:      deploymentName,
+		Namespace: deploymentNamespace,
+	}
+
+	deployment := &appsv1.Deployment{}
+	err := h.KubeRest().Get(context.TODO(), namespacedName, deployment)
+	if err != nil {
+		return &appsv1.Deployment{}, err
+	}
+	return deployment, nil
+}
+
+func (h *SuiteController) DeploymentIsCompleted(deploymentName, namespace string, readyReplicas int32) wait.ConditionFunc {
+	return func() (bool, error) {
+		namespacedName := types.NamespacedName{
+			Name:      deploymentName,
+			Namespace: namespace,
+		}
+
+		deployment := &appsv1.Deployment{}
+		err := h.KubeRest().Get(context.TODO(), namespacedName, deployment)
+		if err != nil && !k8sErrors.IsNotFound(err) {
+			return false, err
+		}
+		if deployment.Status.AvailableReplicas == readyReplicas && deployment.Status.UnavailableReplicas == 0 {
+			return true, nil
+		}
+		return false, nil
+	}
+}
