+++ pkg/utils/common/pod.go
@@ -1,0 +1,109 @@
+package common
+
+import (
+	"bytes"
+	"context"
+	"fmt"
+	"io"
+	"time"
+
+	"github.com/redhat-appstudio/e2e-tests/pkg/utils"
+	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
+	"k8s.io/apimachinery/pkg/util/wait"
+)
+
+func (s *SuiteController) GetPod(namespace, podName string) (*corev1.Pod, error) {
+	return s.KubeInterface().CoreV1().Pods(namespace).Get(context.TODO(), podName, metav1.GetOptions{})
+}
+
+func (s *SuiteController) IsPodRunning(podName, namespace string) wait.ConditionFunc {
+	return func() (bool, error) {
+		pod, err := s.GetPod(namespace, podName)
+		if err != nil {
+			return false, nil
+		}
+		switch pod.Status.Phase {
+		case corev1.PodRunning:
+			return true, nil
+		case corev1.PodFailed, corev1.PodSucceeded:
+			return false, fmt.Errorf("pod %q ran to completion", pod.Name)
+		}
+		return false, nil
+	}
+}
+
+func (s *SuiteController) IsPodSuccessful(podName, namespace string) wait.ConditionFunc {
+	return func() (bool, error) {
+		pod, err := s.GetPod(namespace, podName)
+		if err != nil {
+			return false, nil
+		}
+		switch pod.Status.Phase {
+		case corev1.PodSucceeded:
+			return true, nil
+		case corev1.PodFailed:
+			return false, fmt.Errorf("pod %q has failed", pod.Name)
+		}
+		return false, nil
+	}
+}
+
+func TaskPodExists(tr *v1beta1.TaskRun) wait.ConditionFunc {
+	return func() (bool, error) {
+		if tr.Status.PodName != "" {
+			return true, nil
+		}
+		return false, nil
+	}
+}
+
+func (s *SuiteController) ListPods(namespace, labelKey, labelValue string, selectionLimit int64) (*corev1.PodList, error) {
+	labelSelector := metav1.LabelSelector{MatchLabels: map[string]string{labelKey: labelValue}}
+	listOptions := metav1.ListOptions{
+		LabelSelector: labels.Set(labelSelector.MatchLabels).String(),
+		Limit:         selectionLimit,
+	}
+	return s.KubeInterface().CoreV1().Pods(namespace).List(context.TODO(), listOptions)
+}
+
+func (s *SuiteController) GetContainerLogs(podName, containerName, namespace string) (string, error) {
+	podLogOpts := corev1.PodLogOptions{
+		Container: containerName,
+	}
+
+	req := s.KubeInterface().CoreV1().Pods(namespace).GetLogs(podName, &podLogOpts)
+	podLogs, err := req.Stream(context.TODO())
+	if err != nil {
+		return "", fmt.Errorf("error in opening the stream: %v", err)
+	}
+	defer podLogs.Close()
+
+	buf := new(bytes.Buffer)
+	_, err = io.Copy(buf, podLogs)
+	if err != nil {
+		return "", fmt.Errorf("error in copying logs to buf, %v", err)
+	}
+	return buf.String(), nil
+}
+
+func (s *SuiteController) WaitForPodSelector(
+	fn func(podName, namespace string) wait.ConditionFunc, namespace, labelKey string, labelValue string,
+	timeout int, selectionLimit int64) error {
+	podList, err := s.ListPods(namespace, labelKey, labelValue, selectionLimit)
+	if err != nil {
+		return err
+	}
+	if len(podList.Items) == 0 {
+		return fmt.Errorf("no pods in %s with label key %s and label value %s", namespace, labelKey, labelValue)
+	}
+
+	for i := range podList.Items {
+		if err := utils.WaitUntil(fn(podList.Items[i].Name, namespace), time.Duration(timeout)*time.Second); err != nil {
+			return err
+		}
+	}
+	return nil
+}
